fix timing
(cherry picked from commit ea477bb625f53b03a13a45ad562155419164e45e)

+++ src/main.c
@@ -79,6 +79,6 @@ int main(int argc, char const *argv[]) {
     clock_t start = clock(), diff;
     solve(argv);
     diff = clock() - start;
-    printf("Total time: %lf\n", (double)(diff / CLOCKS_PER_SEC));
+    printf("Total time: %f\n", ((float)(diff) / CLOCKS_PER_SEC));
     return 0;
 }
